+++ src/main/java/walkingkooka/validation/expression/function/ValidationExpressionFunctionValidationErrorIf.java
@@ -15,7 +15,7 @@
  *
  */
 
-package walkingkooka.validation.function;
+package walkingkooka.validation.expression.function;
 
 import walkingkooka.Cast;
 import walkingkooka.collect.set.Sets;
@@ -26,6 +26,7 @@ import walkingkooka.tree.expression.function.ExpressionFunctionParameterKind;
 import walkingkooka.tree.expression.function.ExpressionFunctionParameterName;
 import walkingkooka.validation.ValidationError;
 import walkingkooka.validation.ValidationReference;
+import walkingkooka.validation.expression.ValidatorExpressionEvaluationContext;
 
 import java.util.List;
