+++ src/main/java/walkingkooka/validation/expression/package-info.java
@@ -15,4 +15,4 @@
  *
  */
 
-package walkingkooka.validation.function;
+package walkingkooka.validation.expression;
